remove dataset warning
Dataset not found warning message now pops up whenever habitat values exist but are not nlcd classes. This will happen in a variety of cases causing an unnecessary warning. I'm electing to remove the warning altogether, as I don't know that it is particularly helpful in a site or domain not found scenario either.