Adjustment to match 3.3.8 update
Adjusted code to match update that was applied within Hotfix 3.3.8 to avoid code conflicts when 3.3.8 is merged into biorepo code. Removing underscore from taxon name is needed to avoid taxing the database when illegitimate bots that have been adding the underscore wildcard to taxon names. However, previous change was removing it from all search terms, which proved to be problematic.